[renamerOnUpdate] Refactoring + cleanup
