+++ plugins/renamerOnUpdate/renamerOnUpdate.yml
@@ -1,7 +1,7 @@
 name: renamerOnUpdate
 description: Rename/move filename based on a template.
 url: https://github.com/stashapp/CommunityScripts
-version: 2.4.4
+version: 2.5.0
 exec:
   - python
   - "{pluginDir}/renamerOnUpdate.py"
@@ -12,20 +12,23 @@ hooks:
     triggeredBy:
       - Scene.Update.Post
 tasks:
-  - name: 'Disable'
+  - name: "Disable"
     description: Disable the hook
     defaultArgs:
       mode: disable
-  - name: 'Enable'
+  - name: "Enable"
     description: Enable the hook
     defaultArgs:
       mode: enable
-  - name: 'Dryrun'
+  - name: "Dryrun"
     description: Enable/disable dry-run
     defaultArgs:
       mode: dryrun
-  - name: 'Rename scenes'
+  - name: "Rename scenes"
     description: Rename all your scenes based on your config.
     defaultArgs:
       mode: bulk
-
+  - name: "Rename 15 scenes"
+    description: Rename 15 of your scenes based on your config (useful for testing).
+    defaultArgs:
+      mode: bulk15
